requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
pyTelegramBotAPI==4.15.2